Trait multi_validator_libp2p::infra::RunDA

pub trait RunDA<TYPES, DANET: ConnectedNetwork<Message<TYPES>, TYPES::SignatureKey>, QUORUMNET: ConnectedNetwork<Message<TYPES>, TYPES::SignatureKey>, NODE: NodeImplementation<TYPES, QuorumNetwork = QUORUMNET, CommitteeNetwork = DANET, Storage = TestStorage<TYPES>>>
where
    <TYPES as NodeType>::ValidatedState: TestableState<TYPES>,
    <TYPES as NodeType>::BlockPayload: TestableBlock,
    TYPES: NodeType<Transaction = TestTransaction> + NodeType<InstanceState = TestInstanceState>,
    Leaf<TYPES>: TestableLeaf,
    Self: Sync,
{
    // Required methods
    fn initialize_networking<'async_trait>(
        config: NetworkConfig<TYPES::SignatureKey, TYPES::ElectionConfigType>
    ) -> Pin<Box<dyn Future<Output = Self> + Send + 'async_trait>>
       where Self: 'async_trait;
    fn get_da_channel(&self) -> DANET;
    fn get_quorum_channel(&self) -> QUORUMNET;
    fn get_config(
        &self
    ) -> NetworkConfig<TYPES::SignatureKey, TYPES::ElectionConfigType>;

    // Provided methods
    fn initialize_state_and_hotshot<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = SystemContextHandle<TYPES, NODE>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn run_hotshot<'life0, 'life1, 'async_trait>(
        &'life0 self,
        context: SystemContextHandle<TYPES, NODE>,
        transactions: &'life1 mut Vec<TestTransaction>,
        transactions_to_send_per_round: u64,
        transaction_size_in_bytes: u64
    ) -> Pin<Box<dyn Future<Output = BenchResults> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
}

Defines the behavior of a “run” of the network with a given configuration

Required Methods

fn initialize_networking<'async_trait>( config: NetworkConfig<TYPES::SignatureKey, TYPES::ElectionConfigType> ) -> Pin<Box<dyn Future<Output = Self> + Send + 'async_trait>>
where Self: 'async_trait,

Initializes networking, returns self

fn get_da_channel(&self) -> DANET

Returns the da network for this run

fn get_quorum_channel(&self) -> QUORUMNET

Returns the quorum network for this run

fn get_config( &self ) -> NetworkConfig<TYPES::SignatureKey, TYPES::ElectionConfigType>

Returns the config for this run

Provided Methods

fn initialize_state_and_hotshot<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = SystemContextHandle<TYPES, NODE>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Initializes the genesis state and HotShot instance; does not start HotShot consensus

Panics if it cannot generate a genesis block, fails to initialize HotShot, or cannot

get the anchored view Note: sequencing leaf does not have state, so does not return state

fn run_hotshot<'life0, 'life1, 'async_trait>( &'life0 self, context: SystemContextHandle<TYPES, NODE>, transactions: &'life1 mut Vec<TestTransaction>, transactions_to_send_per_round: u64, transaction_size_in_bytes: u64 ) -> Pin<Box<dyn Future<Output = BenchResults> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Starts HotShot consensus, returns when consensus has finished

Object Safety

This trait is not object safe.